add big changes in all files

+++ client.js
@@ -1,43 +1,53 @@
 // client.js
 const net = require('net');
-const fs = require('fs');   // добавление fs
+const fs = require('fs');
 const port = 8124;
+const string = 'QA';
+const bad = 'DEC';
+const good = 'ACK';
 
 const client = new net.Socket();
-
+let currentIndex = -1;
 client.setEncoding('utf8');
-let arrquery;
 
-client.connect(port, function() {
-    console.log('Connected');
-    fs.readFile('qa.json', (e,text) => {
-
-        if(e) {
-            console.log(e)
+let questions = [];
+client.connect({port: port, host: '127.0.0.1'}, () => {
+    fs.readFile("qa.json", (err, text) => {
+        if (!err) {
+            questions = JSON.parse(text);
+            client.write(string);
         }
-        else {
-            arrquery = sh(JSON.parse(text));
-            client.write('QA');
-        }
+        else console.error(err);
     });
 });
 
-client.on('data', function(data) {
-    console.log(data);
-    client.destroy();
+client.on('data', (data) => {
+    if (data === bad)
+        client.destroy();
+    if (data === good)
+        sendQuestion();
+    else {
+        let qst = questions[currentIndex];
+        let answer = qst.good;
+        console.log('\n' + qst.quest);
+        console.log('Answer:' + data);
+        console.log('Server:' + answer);
+        console.log('Result:' + (data === answer ? 'True': 'False'));
+        sendQuestion();
+    }
 });
 
-client.on('close', function() {
+client.on('close', function () {
     console.log('Connection closed');
 });
 
-function sh(array) {
-    let a;
-    var result = [];
-    while(array.length > 0){
-        a = Math.floor(Math.random() * array.length);
-        result.push(array[a]);
-        array.splice(a, 1);
+
+
+function sendQuestion() {
+    if (currentIndex < questions.length -1) {
+        let qst = questions[++currentIndex].quest;
+        client.write(qst);
     }
-    return result;
+    else
+        client.destroy();
 }
